a few file renames
