+++ src/NoSidebarRouter.js
@@ -4,8 +4,7 @@ import ArticlesRouter from "./articles/_ArticlesRouter";
 import WordsRouter from "./words/_WordsRouter";
 import ExercisesRouter from "./exercises/ExercisesRouter";
 import TeacherRouter from "./teacher/_routing/_TeacherRouter";
-import Settings from "./pages/settings_pages/Settings";
-import ArticleReader from "./reader/ArticleReader";
+import Settings from "./pages/Settings/Settings";
 import UserDashboard from "./userDashboard/UserDashboard";
 import React from "react";
 import ReadingHistory from "./words/WordHistory";
