+++ src/articles/ExtensionMessage.js
@@ -3,7 +3,7 @@ import Header from "../components/modal_shared/Header";
 import Heading from "../components/modal_shared/Heading";
 import Main from "../components/modal_shared/Main";
 import FullWidthImage from "../components/FullWidthImage";
-import Footer from "../pages/info_page_shared/Footer";
+import Footer from "../pages/_pages_shared/Footer";
 import ButtonContainer from "../components/modal_shared/ButtonContainer";
 import GoToButton from "../components/modal_shared/GoToButton";
 import { getExtensionInstallationLinks } from "../utils/extension/extensionInstallationLinks";
