+++ src/landingPage/LandingPage.js
@@ -5,7 +5,7 @@ import Contributors from "./Contributors";
 import { Redirect } from "react-router-dom";
 import { setTitle } from "../assorted/setTitle";
 import { getSessionFromCookies } from "../utils/cookies/userInfo";
-import Button from "../pages/info_page_shared/Button.js";
+import Button from "../pages/_pages_shared/Button.js";
 import ArrowForwardRoundedIcon from "@mui/icons-material/ArrowForwardRounded";
 
 import redirect from "../utils/routing/routing.js";
