+++ src/pages/CreateAccount.js
@@ -7,17 +7,17 @@ import useFormField from "../hooks/useFormField";
 import { UserContext } from "../contexts/UserContext";
 import { saveUserInfoIntoCookies } from "../utils/cookies/userInfo";
 
-import InfoPage from "./info_page_shared/InfoPage";
-import Header from "./info_page_shared/Header";
-import Heading from "./info_page_shared/Heading";
-import Main from "./info_page_shared/Main";
-import Form from "./info_page_shared/Form";
-import FullWidthErrorMsg from "./info_page_shared/FullWidthErrorMsg";
-import FormSection from "./info_page_shared/FormSection";
+import InfoPage from "./_pages_shared/InfoPage";
+import Header from "./_pages_shared/Header";
+import Heading from "./_pages_shared/Heading";
+import Main from "./_pages_shared/Main";
+import Form from "./_pages_shared/Form";
+import FullWidthErrorMsg from "./_pages_shared/FullWidthErrorMsg";
+import FormSection from "./_pages_shared/FormSection";
 import InputField from "../components/InputField";
-import Footer from "./info_page_shared/Footer";
-import ButtonContainer from "./info_page_shared/ButtonContainer";
-import Button from "./info_page_shared/Button";
+import Footer from "./_pages_shared/Footer";
+import ButtonContainer from "./_pages_shared/ButtonContainer";
+import Button from "./_pages_shared/Button";
 import Modal from "../components/modal_shared/Modal";
 import validator from "../assorted/validator";
 import strings from "../i18n/definitions";
