+++ src/pages/ExcludeWordsStep2.js
@@ -3,21 +3,21 @@ import { isSupportedBrowser } from "../utils/misc/browserDetection";
 import useUnwantedContentPreferences from "../hooks/useUnwantedContentPreferences";
 import useFormField from "../hooks/useFormField";
 
-import InfoPage from "./info_page_shared/InfoPage";
-import Header from "./info_page_shared/Header";
-import Heading from "./info_page_shared/Heading";
-import Main from "./info_page_shared/Main";
-import Form from "./info_page_shared/Form";
-import Button from "./info_page_shared/Button";
-import Footer from "./info_page_shared/Footer";
-import ButtonContainer from "./info_page_shared/ButtonContainer";
+import InfoPage from "./_pages_shared/InfoPage";
+import Header from "./_pages_shared/Header";
+import Heading from "./_pages_shared/Heading";
+import Main from "./_pages_shared/Main";
+import Form from "./_pages_shared/Form";
+import Button from "./_pages_shared/Button";
+import Footer from "./_pages_shared/Footer";
+import ButtonContainer from "./_pages_shared/ButtonContainer";
 import ArrowForwardRoundedIcon from "@mui/icons-material/ArrowForwardRounded";
 import AddRoundedIcon from "@mui/icons-material/AddRounded";
 import HighlightOffRoundedIcon from "@mui/icons-material/HighlightOffRounded";
 
 import InputField from "../components/InputField";
-import Tag from "./info_page_shared/Tag";
-import TagContainer from "./info_page_shared/TagContainer";
+import Tag from "./_pages_shared/Tag";
+import TagContainer from "./_pages_shared/TagContainer";
 
 import redirect from "../utils/routing/routing";
 import strings from "../i18n/definitions";
