+++ src/pages/InstallExtension.js
@@ -1,10 +1,10 @@
-import InfoPage from "./info_page_shared/InfoPage";
-import Header from "./info_page_shared/Header";
-import Heading from "./info_page_shared/Heading";
-import Main from "./info_page_shared/Main";
-import ButtonContainer from "./info_page_shared/ButtonContainer";
-import Footer from "./info_page_shared/Footer";
-import Button from "./info_page_shared/Button";
+import InfoPage from "./_pages_shared/InfoPage";
+import Header from "./_pages_shared/Header";
+import Heading from "./_pages_shared/Heading";
+import Main from "./_pages_shared/Main";
+import ButtonContainer from "./_pages_shared/ButtonContainer";
+import Footer from "./_pages_shared/Footer";
+import Button from "./_pages_shared/Button";
 import FullWidthImage from "../components/FullWidthImage";
 import { getExtensionInstallationLinks } from "../utils/extension/extensionInstallationLinks";
