+++ src/pages/LanguagePreferences.js
@@ -5,16 +5,16 @@ import LocalStorage from "../assorted/LocalStorage";
 import redirect from "../utils/routing/routing";
 import useFormField from "../hooks/useFormField";
 
-import InfoPage from "./info_page_shared/InfoPage";
-import Header from "./info_page_shared/Header";
-import Heading from "./info_page_shared/Heading";
-import Main from "./info_page_shared/Main";
-import Form from "./info_page_shared/Form";
-import FullWidthErrorMsg from "./info_page_shared/FullWidthErrorMsg";
-import FormSection from "./info_page_shared/FormSection";
+import InfoPage from "./_pages_shared/InfoPage";
+import Header from "./_pages_shared/Header";
+import Heading from "./_pages_shared/Heading";
+import Main from "./_pages_shared/Main";
+import Form from "./_pages_shared/Form";
+import FullWidthErrorMsg from "./_pages_shared/FullWidthErrorMsg";
+import FormSection from "./_pages_shared/FormSection";
 import Selector from "../components/Selector";
-import ButtonContainer from "./info_page_shared/ButtonContainer";
-import Button from "./info_page_shared/Button";
+import ButtonContainer from "./_pages_shared/ButtonContainer";
+import Button from "./_pages_shared/Button";
 import ArrowForwardRoundedIcon from "@mui/icons-material/ArrowForwardRounded";
 
 import validator from "../assorted/validator";
