+++ src/pages/ResetPasswordStep2.js
@@ -4,12 +4,12 @@ import strings from "../i18n/definitions";
 
 import useFormField from "../hooks/useFormField";
 
-import Form from "./info_page_shared/Form";
-import FormSection from "./info_page_shared/FormSection";
-import FullWidthErrorMsg from "./info_page_shared/FullWidthErrorMsg";
+import Form from "./_pages_shared/Form";
+import FormSection from "./_pages_shared/FormSection";
+import FullWidthErrorMsg from "./_pages_shared/FullWidthErrorMsg";
 import InputField from "../components/InputField";
-import ButtonContainer from "./info_page_shared/ButtonContainer";
-import Button from "./info_page_shared/Button";
+import ButtonContainer from "./_pages_shared/ButtonContainer";
+import Button from "./_pages_shared/Button";
 
 export default function ResetPasswordStep2({ api, email }) {
   const [errorMessage, setErrorMessage] = useState("");
