+++ src/pages/SelectInterests.js
@@ -1,13 +1,13 @@
 import useSelectInterest from "../hooks/useSelectInterest";
-import InfoPage from "./info_page_shared/InfoPage";
-import Header from "./info_page_shared/Header";
-import Heading from "./info_page_shared/Heading";
-import Main from "./info_page_shared/Main";
-import ButtonContainer from "./info_page_shared/ButtonContainer";
-import Footer from "./info_page_shared/Footer";
-import Button from "./info_page_shared/Button";
-import Tag from "./info_page_shared/Tag";
-import TagContainer from "./info_page_shared/TagContainer";
+import InfoPage from "./_pages_shared/InfoPage";
+import Header from "./_pages_shared/Header";
+import Heading from "./_pages_shared/Heading";
+import Main from "./_pages_shared/Main";
+import ButtonContainer from "./_pages_shared/ButtonContainer";
+import Footer from "./_pages_shared/Footer";
+import Button from "./_pages_shared/Button";
+import Tag from "./_pages_shared/Tag";
+import TagContainer from "./_pages_shared/TagContainer";
 import ArrowForwardRoundedIcon from "@mui/icons-material/ArrowForwardRounded";
 import strings from "../i18n/definitions";
