+++ src/pages/Settings/AudioExercises.js
@@ -4,14 +4,14 @@ import { UserContext } from "../../contexts/UserContext";
 import SessionStorage from "../../assorted/SessionStorage";
 import LocalStorage from "../../assorted/LocalStorage";
 import Feature from "../../features/Feature";
-import Button from "../info_page_shared/Button";
-import ButtonContainer from "../info_page_shared/ButtonContainer";
-import Form from "../info_page_shared/Form";
-import FormSection from "../info_page_shared/FormSection";
-import InfoPage from "../info_page_shared/InfoPage";
-import Main from "../info_page_shared/Main";
-import Header from "../info_page_shared/Header";
-import Heading from "../info_page_shared/Heading";
+import Button from "../_pages_shared/Button";
+import ButtonContainer from "../_pages_shared/ButtonContainer";
+import Form from "../_pages_shared/Form";
+import FormSection from "../_pages_shared/FormSection";
+import InfoPage from "../_pages_shared/InfoPage";
+import Main from "../_pages_shared/Main";
+import Header from "../_pages_shared/Header";
+import Heading from "../_pages_shared/Heading";
 import BackArrow from "./settings_pages_shared/BackArrow";
 import Checkbox from "../../components/modal_shared/Checkbox";
 import strings from "../../i18n/definitions";
