+++ src/pages/Settings/CurrentClass.js
@@ -1,16 +1,16 @@
 import { useHistory } from "react-router-dom";
 import { useState, useEffect, useContext } from "react";
 import { UserContext } from "../../contexts/UserContext";
-import Button from "../info_page_shared/Button";
-import ButtonContainer from "../info_page_shared/ButtonContainer";
+import Button from "../_pages_shared/Button";
+import ButtonContainer from "../_pages_shared/ButtonContainer";
 import InputField from "../../components/InputField";
-import Form from "../info_page_shared/Form";
-import FormSection from "../info_page_shared/FormSection";
-import InfoPage from "../info_page_shared/InfoPage";
-import Header from "../info_page_shared/Header";
-import Heading from "../info_page_shared/Heading";
-import Main from "../info_page_shared/Main";
-import FullWidthErrorMsg from "../info_page_shared/FullWidthErrorMsg";
+import Form from "../_pages_shared/Form";
+import FormSection from "../_pages_shared/FormSection";
+import InfoPage from "../_pages_shared/InfoPage";
+import Header from "../_pages_shared/Header";
+import Heading from "../_pages_shared/Heading";
+import Main from "../_pages_shared/Main";
+import FullWidthErrorMsg from "../_pages_shared/FullWidthErrorMsg";
 import BackArrow from "./settings_pages_shared/BackArrow";
 import strings from "../../i18n/definitions";
 
@@ -51,6 +51,7 @@ export default function CurrentClass({ api }) {
       },
     );
   }
+
   return (
     <InfoPage layoutVariant={"minimalistic-top-aligned"}>
       <BackArrow />
