+++ src/pages/Settings/Languages.js
@@ -7,15 +7,15 @@ import strings from "../../i18n/definitions";
 import LocalStorage from "../../assorted/LocalStorage";
 import LoadingAnimation from "../../components/LoadingAnimation";
 import LanguageSelector from "../../components/LanguageSelector";
-import Button from "../info_page_shared/Button";
-import ButtonContainer from "../info_page_shared/ButtonContainer";
-import Form from "../info_page_shared/Form";
-import FormSection from "../info_page_shared/FormSection";
-import InfoPage from "../info_page_shared/InfoPage";
-import Header from "../info_page_shared/Header";
-import Heading from "../info_page_shared/Heading";
-import Main from "../info_page_shared/Main";
-import FullWidthErrorMsg from "../info_page_shared/FullWidthErrorMsg";
+import Button from "../_pages_shared/Button";
+import ButtonContainer from "../_pages_shared/ButtonContainer";
+import Form from "../_pages_shared/Form";
+import FormSection from "../_pages_shared/FormSection";
+import InfoPage from "../_pages_shared/InfoPage";
+import Header from "../_pages_shared/Header";
+import Heading from "../_pages_shared/Heading";
+import Main from "../_pages_shared/Main";
+import FullWidthErrorMsg from "../_pages_shared/FullWidthErrorMsg";
 import BackArrow from "./settings_pages_shared/BackArrow";
 import Selector from "../../components/Selector";
