+++ src/pages/_pages_shared/Form.sc.js
@@ -3,7 +3,6 @@ import styled from "styled-components";
 const Form = styled.form`
   display: flex;
   flex-direction: column;
-  align-items: stretch;
   align-items: center;
   gap: 1.5rem;
   width: 100%;
