+++ src/pages/_pages_shared/Logo.js
